pyproject.toml
[project]
name = "guitar-pro-tab-tune-organizer"
version = "0.0.1"
dependencies = [
    "PyGuitarPro~=0.9.2",
]
readme = "README.md"
requires-python = ">=3.7"

[project.scripts]
guitar-pro-organize-tabs = "organize_tabs:cli"

[project.optional-dependencies]
tests = ["pytest"]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = ["organize_tabs"]
